requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.25
django-rest-swagger==2.2.0
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
pkg-resources==0.0.0
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.1
requests==2.27.1
simplejson==3.19.2
sqlparse==0.4.4
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.18
uuid==1.30
zipp==3.6.0
